Reintroduced Clean plugin. Refactored Mac and iOS Compile plugin
The Clean plugin required updating to the 'cocos' namespace. The
execution dependency on dist was removed (since that's an optional
plugin). The Clean now does the clean with xcodebuild so there's
a connection with the Compile plugin class to access the same
build environment variables.

Tested compile, run, and clean commands on Mac and iOS